migration: Migrate data from edit_history to edit_history_entries.
We cast edit_history because it already stores valid JSON consistently.
Skipping the null values because we want to fall back to "[]" in that case.

We also implement double-write for edit_history_entries so that we have a
window to run the backfilling migration without downtime or extensive
locking, which also prepares us to stop using the edit_history field.

Co-Authored-By: Zixuan James Li <p359101898@gmail.com>