fixup! migration: Migrate data from extra_data to extra_data_json.
(squash) migration: Check for inconsistent audit log entries.

This prints out a warning when the extra_data_json field of an entry
would have been overwritten by a value inconsistent with what we derived
from extra_data. Usually this only happens when the extra_data was
corrupted before this migration.