user_groups: Test user group race conditions.
This implements a transaction test case targeting some typical scenarios
when an internal server error is expected to happen (this means that the
user group view makes the correct decision to abort the transaction when
something goes wrong with locks).

To achieve this, we add a development view intended only for unit tests.
It has a global BARRIER that can be shared across threads, so that we
can synchronize them to consistently reproduce certain potential race
conditions prevented by the database locks.

The transaction test case lanuches pairs of threads initiating
conflicting requests at the same time. The tests are set up such that
exactly one of them is expected to succeed with a certain error message
(while we don't know each one).